KA-269 Rename memorySelector to .ts, install types
Types for mem and immutable
