+++ src/selectors/memorySelector.ts
